requirements.txt
telethon==1.32.1
Pillow>=9.0.0
requests>=2.25.0
openai==1.12.0
pywin32==306
python-dotenv==1.0.1
cryptg==0.4.0
aiohttp==3.9.3
packaging>=21.0
psutil>=5.8.0
send2trash>=1.8.0
winshell>=0.6 